Added random skin loading, changed duration to be long;

+++ BackgroundPlayer/Configuration/SkinConfig.cs
@@ -4,7 +4,7 @@ namespace BackgroundPlayer.Configuration
 {
     public class SkinConfig
     {
-        public int DurationMillisecods { get; set; }
+        public long DurationMillisecods { get; set; }
 
         public StartOffset StartOffset { get; set; }
     }
